BestSorter.py
I ran into some issues with pathfile character limit in windows. Didn't help that it gave me a FileNotFound error, so it took me a while to figure out what was wrong. 
Folderpaths over 200 characters will be shortened, but the first 6 header columns will be preserved, along with the last column.